feat(payment): PAYPAL-00 moved ratepay strategy to separate directory
